Refactored code to make it easier to support databases from Worldbirds.
Changed the way the record is loaded from the file to be converted so that
only the required headers are used. This avoids the problem where different
databases from BirdLife International's Worldbirds series have slighting
different record formats.

Removed the locations table that is used to map locations from the source
database to locations in eBird. Maintaining the data is very time-consuming
and it is easier to let the person importing the records into eBird to
manange the mapping as part of the import process. This is less efficient
but it gives greater flexibility in deciding which locations should be
merged into existing eBird hotspots.

Updated the way the handlers resources are intialized to make it easier to
configure resources (currently only species names) on a per database basis.