Restructured package.
Reoranised the layout of the library so everything was not part of the source
tree.

Renamed the package to ebird_import to make the purpose a little cleaer.

Removed the dependency on nose. The test run just fine with the unittest runner.

Fixed the tests.

Updated the data files used for species name translation from worldbirds to ebird.